ghah.  got more tests working properly, but icky gross
the tests do not work at all if I drop down to visitor user because my
sql fu is lame and I can't figure out a way to tell the visitor user
the correct invitation_id if it cannot look at the invitations table.
I need to somehow stash that invite id value in a table or variable or
something, but whatever.  Test is mostly correct running as superuser,
so its fine, its fine.